// angular/src/gs-api/src/model/authenticationRequest.ts
/**
 * Gestion de stock REST API documentation
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthenticationRequest { 
    login?: string;
    password?: string;
}